+++ src/Service/Kms/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Added `me-central-1`, `me-central-1-fips`, `ap-south-2-fips`, `eu-south-2-fips`, `eu-central-2-fips`, `eu-central-2`, `eu-south-2` and `ap-south-2` regions
+- AWS api-change: AWS KMS introduces the External Key Store (XKS), a new feature for customers who want to protect their data with encryption keys stored in an external key management system under their control.
 
 ## 0.1.1